Fix using keyboard shortcuts when in a full-screen app
Third-party apps can only control HomeKit when they're “in the foreground”. On macOS, this just means having a piece of UI visible, such as the menubar icon.
As a result, when an app is in full-screen mode, HomeConMenu keyboard shortcuts stop working.

This commit creates an ever-present invisible window, which fixes the issue.